+++ examples/example-app-v12-monorepo/projects/app1/jest-esm.config.mjs
@@ -1,8 +1,5 @@
-const { pathsToModuleNameMapper } = require('ts-jest/utils');
-const { paths } = require('./tsconfig.json').compilerOptions;
-
 /** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
-module.exports = {
+const jestConfig = {
   preset: 'jest-preset-angular/presets/defaults-esm',
   globals: {
     'ts-jest': {
@@ -12,8 +9,9 @@ module.exports = {
     },
   },
   moduleNameMapper: {
-    ...pathsToModuleNameMapper(paths, { prefix: '<rootDir>' }),
     tslib: 'tslib/tslib.es6.js',
   },
   setupFilesAfterEnv: ['<rootDir>/setup-jest.ts'],
-};
+}
+
+export default jestConfig;