+++ examples/example-app-v13/jest-esm.config.mjs
@@ -1,5 +1,5 @@
 /** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
-module.exports = {
+const jestConfig = {
   preset: 'jest-preset-angular/presets/defaults-esm',
   globals: {
     'ts-jest': {
@@ -8,9 +8,12 @@ module.exports = {
       tsconfig: '<rootDir>/tsconfig-esm.spec.json',
     },
   },
+  globalSetup: 'jest-preset-angular/global-setup',
   moduleNameMapper: {
-    '^@app2/services/(.*)$': '<rootDir>/src/app/services/$1',
     tslib: 'tslib/tslib.es6.js',
+    rxjs: '<rootDir>/node_modules/rxjs/dist/bundles/rxjs.umd.js'
   },
   setupFilesAfterEnv: ['<rootDir>/setup-jest.ts'],
-};
+}
+
+export default jestConfig;